+++ src/ConfigOptions/DbCredentialsFileSettings.cs
@@ -10,6 +10,5 @@ public class DbCredentialsFileSettings
     public string Database { get; set; }
     public int Port { get; set; }
     public string ApplicationName { get; set; }
-    public bool Pooling { get; set; }
-    public bool TrustServerCertificate { get; set; }
+    public string AdditionalParameters { get; set; }
 }